Add Play Setup and Live Match Shells with Lifecycle Management
- Introduced `PlaySetupShell.vue` and `LiveMatchShell.vue` components for managing game setup and live match states.
- Implemented lifecycle management for the live match shell in `liveMatchShellLifecycle.js`, including activation and deactivation logic.
- Added test cases for new components and lifecycle functions to ensure proper functionality and integration.
- Created supporting files for test IDs and shell activation logic to enhance maintainability and clarity in the codebase.
