+++ src/features/dungeon-runner/ui/shells/liveMatchShellSessionKey.js
@@ -1,0 +1,1 @@
+export const LIVE_MATCH_SHELL_SESSION_KEY = Symbol('liveMatchShellSession')
