+++ src/features/dungeon-runner/ui/shells/liveMatchShellTestIds.js
@@ -1,0 +1,6 @@
+export const LIVE_MATCH_SHELL_TEST_IDS = {
+  root: 'live-match-shell',
+  finishingMatchOverlay: 'finishing-match-overlay',
+  neuralRefreshTerminal: 'neural-refresh-terminal',
+  neuralRefreshTerminalReload: 'neural-refresh-terminal-reload',
+}
