+++ src/features/dungeon-runner/ui/shells/matchOverShellTestIds.js
@@ -1,0 +1,6 @@
+export const MATCH_OVER_SHELL_TEST_IDS = {
+  root: 'match-over-shell',
+  rematch: 'match-over-rematch',
+  backToSetup: 'match-over-back-to-setup',
+  exportReplay: 'match-over-export-replay',
+}
